+++ src/main/objhits.c
@@ -1717,7 +1717,7 @@ void ObjHits_ApplyPairResponse(int objA, int objB, f32 x, f32 y, f32 z, int flag
     *(int*)stateB = objA;
     if (animA->parent != NULL)
     {
-        Obj_TransformWorldVectorToLocal(x, y, z, &localAx, &localAy, &localAz, (int)animA->parent);
+        Obj_TransformWorldVectorToLocal(x, y, z, &localAx, &localAy, &localAz, animA->parent);
     }
     else
     {
@@ -1727,7 +1727,7 @@ void ObjHits_ApplyPairResponse(int objA, int objB, f32 x, f32 y, f32 z, int flag
     }
     if (animB->parent != NULL)
     {
-        Obj_TransformWorldVectorToLocal(x, y, z, &localBx, &localBy, &localBz, (int)animB->parent);
+        Obj_TransformWorldVectorToLocal(x, y, z, &localBx, &localBy, &localBz, animB->parent);
     }
     else
     {
@@ -1750,7 +1750,7 @@ void ObjHits_ApplyPairResponse(int objA, int objB, f32 x, f32 y, f32 z, int flag
         else
         {
             Obj_TransformLocalPointToWorld(animA->localPosX, animA->localPosY, animA->localPosZ, &animA->worldPosX,
-                                           &animA->worldPosY, &animA->worldPosZ, (int)animA->parent);
+                                           &animA->worldPosY, &animA->worldPosZ, animA->parent);
         }
     }
     else if ((animB->classId == 1) && (stateB->lateralResponseWeight != 0) &&
@@ -1768,7 +1768,7 @@ void ObjHits_ApplyPairResponse(int objA, int objB, f32 x, f32 y, f32 z, int flag
         else
         {
             Obj_TransformLocalPointToWorld(animB->localPosX, animB->localPosY, animB->localPosZ, &animB->worldPosX,
-                                           &animB->worldPosY, &animB->worldPosZ, (int)animB->parent);
+                                           &animB->worldPosY, &animB->worldPosZ, animB->parent);
         }
     }
     else if (stateB->lateralResponseWeight == 0)
@@ -1787,7 +1787,7 @@ void ObjHits_ApplyPairResponse(int objA, int objB, f32 x, f32 y, f32 z, int flag
             else
             {
                 Obj_TransformLocalPointToWorld(animA->localPosX, animA->localPosY, animA->localPosZ, &animA->worldPosX,
-                                               &animA->worldPosY, &animA->worldPosZ, (int)animA->parent);
+                                               &animA->worldPosY, &animA->worldPosZ, animA->parent);
             }
         }
     }
@@ -1807,7 +1807,7 @@ void ObjHits_ApplyPairResponse(int objA, int objB, f32 x, f32 y, f32 z, int flag
             else
             {
                 Obj_TransformLocalPointToWorld(animB->localPosX, animB->localPosY, animB->localPosZ, &animB->worldPosX,
-                                               &animB->worldPosY, &animB->worldPosZ, (int)animB->parent);
+                                               &animB->worldPosY, &animB->worldPosZ, animB->parent);
             }
         }
     }
@@ -1859,13 +1859,13 @@ void ObjHits_ApplyPairResponse(int objA, int objB, f32 x, f32 y, f32 z, int flag
         animA->localPosY = animA->localPosY - localAy * blend;
         animA->localPosZ = animA->localPosZ - localAz * blend;
         Obj_TransformLocalPointToWorld(animA->localPosX, animA->localPosY, animA->localPosZ, &animA->worldPosX,
-                                       &animA->worldPosY, &animA->worldPosZ, (int)animA->parent);
+                                       &animA->worldPosY, &animA->worldPosZ, animA->parent);
         invBlend = gObjHitsScalarOne - blend;
         animB->localPosX = localBx * invBlend + animB->localPosX;
         animB->localPosY = localBy * invBlend + animB->localPosY;
         animB->localPosZ = localBz * invBlend + animB->localPosZ;
         Obj_TransformLocalPointToWorld(animB->localPosX, animB->localPosY, animB->localPosZ, &animB->worldPosX,
-                                       &animB->worldPosY, &animB->worldPosZ, (int)animB->parent);
+                                       &animB->worldPosY, &animB->worldPosZ, animB->parent);
     }
 }
 
@@ -2545,7 +2545,7 @@ void ObjHits_Update(int objectCount)
         {
             Obj_TransformLocalPointToWorld(objState->localPosX, objState->localPosY, objState->localPosZ,
                                            &objState->worldPosX, &objState->worldPosY, &objState->worldPosZ,
-                                           (int)((GameObject*)obj)->anim.parent);
+                                           ((GameObject*)obj)->anim.parent);
         }
         else
         {
